+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 UpTab adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [V1.2.1] — 2024-07-26
+
+### ✨ Added
+
+- A new background: a photograph of the "Lage Zwaluwe - 's Hertogenbosch" railway. Thanks for the contribution, Max!
+- Two more new backgrounds: Tivoli Vredenburg in Utrecht and a bad-ass Dutch Castle
+
+### ♻️ Changed
+
+- Various dependency updates
+
 ## [V1.2.0] — 2024-02-16
 
 ### ✨ Added
@@ -75,6 +86,7 @@ UpTab adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - Added Bookmarks: allowing you to easily view, open and delete your browser bookmarks
   - Added `CTRL+U` shortkey to open popup
 
+[V1.2.1]: https://github.com/AJGeel/uptab/compare/v1.2.1...v1.2.0
 [V1.2.0]: https://github.com/AJGeel/uptab/compare/v1.2.0...v1.1.0
 [V1.1.0]: https://github.com/AJGeel/uptab/compare/v1.1.0...v1.0.1
 [V1.0.1]: https://github.com/AJGeel/uptab/compare/v1.0.0...v1.0.1